WIP: Localize menu items, error messages
Detecting runtime changes to the GUI language is not trivial and
may not even be possible, given the lack of dedicated APIs and the
limitations on window procedures in a plugin's constructor, when the
editor's handle in not yet available

Sending NPPM_GETPLUGINSCONFIGDIR fails without a handle, making a
configurable option for menu titles impractical. Translations will
be saved only with the DLL for now. A 'nativeLang.xml' must be found
in the appropriate config directory